Replace fragment listeners by otto events.
Listeners are fine, but since we use Otto already we might as well
leverage it and remove the listeners. Listeners are always troublesome
because they are hard to handle (i.e. most listeners won't unregister
themselves leaving references to potentially inconsistent resources) and
writing good listener management is not trivial. The event bus solves
this for us.

Since Otto is now used everywhere, I also added a OpenSpritzApplication
that contains Otto and might contain other global stuff in the future.