Article Sharing Changes
Always submit article reads to the Glance network for articles from
within the Glance network. e.g: If you click a story in the Digg (or
any built-in) Feed, always increment the read account for that article
and add it to the Recents and Popular Feed if appropriate.

Allow users the option to configure this behavior for articles shared
to Glance from external applications (e.g: the user’s web browser).
This option is still available under the Settings menu, but now
defaults to “ask” instead of “never”